refactor(reference-service)!: rename service layer to services
Directory `service/` becomes `services/` (singular to plural), following
the same pattern as the prior application-to-service rename. Import
paths, the mypy strict-override module list, and the import-linter
contract (source_modules, forbidden_modules, and its name, kept
grammatical: "Services import neither api nor infrastructure") all
follow the new path. No behaviour, tests, or assertions change.
