+++ examples/reference-service/src/reference_service/api/deps.py
@@ -12,7 +12,7 @@ from fastapi import Depends, Request
 
 from reference_service.container import Container
 from reference_service.domain.repositories import OrderRepository
-from reference_service.service.order import GetOrder, PlaceOrder
+from reference_service.services.order import GetOrder, PlaceOrder
 
 
 def get_container(request: Request) -> Container:
